+++ Cesium3DTilesSelection/src/GltfContent.cpp
@@ -2,9 +2,11 @@
 #include "Cesium3DTilesSelection/spdlog-cesium.h"
 #include "CesiumGltf/AccessorView.h"
 #include "CesiumGltf/AccessorWriter.h"
+#include "CesiumGltf/Model.h"
 #include "CesiumUtility/Math.h"
 #include "CesiumUtility/Tracing.h"
 #include "CesiumUtility/joinToString.h"
+#include <optional>
 #include <stdexcept>
 
 namespace Cesium3DTilesSelection {
@@ -206,8 +208,13 @@ GltfContent::createRasterOverlayTextureCoordinates(
   double minimumHeight = std::numeric_limits<double>::max();
   double maximumHeight = std::numeric_limits<double>::lowest();
 
-  Gltf::forEachPrimitiveInScene(
-      gltf,
+  static glm::dmat4 gltfToCesiumAxes(
+      glm::dvec4(1.0, 0.0, 0.0, 0.0),
+      glm::dvec4(0.0, 0.0, 1.0, 0.0),
+      glm::dvec4(0.0, -1.0, 0.0, 0.0),
+      glm::dvec4(0.0, 0.0, 0.0, 1.0));
+
+  gltf.forEachPrimitiveInScene(
       -1,
       [&transform,
        &positionAccessorsToTextureCoordinateAccessor,
@@ -250,6 +257,8 @@ GltfContent::createRasterOverlayTextureCoordinates(
           return;
         }
 
+        glm::dmat4 transform = gltfToCesiumAxes * nodeTransform;
+
         // Generate new texture coordinates
         int nextTextureCoordinateAccessorIndex =
             generateOverlayTextureCoordinates(
@@ -279,5 +288,4 @@ GltfContent::createRasterOverlayTextureCoordinates(
       minimumHeight,
       maximumHeight);
 }
-
 } // namespace Cesium3DTilesSelection