+++ Cesium3DTilesSelection/src/Tile.cpp
@@ -9,6 +9,7 @@
 #include "CesiumGeometry/Axis.h"
 #include "CesiumGeometry/AxisTransforms.h"
 #include "CesiumGeospatial/Transforms.h"
+#include "CesiumGltf/Model.h"
 #include "CesiumUtility/Tracing.h"
 #include "TileUtilities.h"
 #include "upsampleGltfForRasterOverlays.h"
@@ -266,7 +267,10 @@ void Tile::loadContent() {
            gltfUpAxis,
            pPrepareRendererResources =
                tileset.getExternals().pPrepareRendererResources,
-           pLogger = tileset.getExternals().pLogger](
+           pLogger = tileset.getExternals().pLogger,
+           generateMissingNormalsSmooth =
+               tileset.getOptions()
+                   .contentOptions.generateMissingNormalsSmooth](
               std::shared_ptr<IAssetRequest>&& pRequest) mutable {
             CESIUM_TRACE("loadContent worker thread");
             const IAssetResponse* pResponse = pRequest->response();
@@ -329,6 +333,10 @@ void Tile::loadContent() {
                     transform,
                     boundingVolume,
                     projections);
+
+                if (generateMissingNormalsSmooth) {
+                  pContent->model->generateMissingNormalsSmooth();
+                }
 
                 if (pPrepareRendererResources) {
                   CESIUM_TRACE("prepareInLoadThread");
@@ -859,6 +867,7 @@ void Tile::upsampleParent(
             std::make_unique<TileContentLoadResult>();
         pContent->model =
             upsampleGltfForRasterOverlays(parentModel, *pSubdividedParentID);
+
         if (pContent->model) {
           pContent->updatedBoundingVolume = Tile::generateTextureCoordinates(
               pContent->model.value(),