+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 ##### Breaking Changes :mega:
 
+- Deleted `Cesium3DTilesSelection::Gltf` and moved functionality into `CesiumGltf::Model`. 
 - Renamed `Rectangle::intersect` and `GlobeRectangle::intersect` to `computeIntersection`.
 - `RasterOverlay` and derived classes now require a `name` parameter to their constructors.
 - Changed the type of texture coordinate IDs used in the raster overlay system from `uint32_t` to `int32_t`.
@@ -14,6 +15,7 @@
 
 - Added `Future<T>::isReady`.
 - Added `Future<T>::share`, which returns a `SharedFuture<T>` and allows multiple continuations to be attached.
+- Added an option in `TilesetOptions::ContentOptions` to generate smooth normals when the original glTFs were missing normals.
 - Added `ImageManipulation` class to `CesiumGltfReader`.
 - Added `Math::roundUp` and `Math::roundDown`.
 - Added `Rectangle::computeUnion`.